cpp version fixed and tested

+++ cpp/NSKalgo.h
@@ -7,19 +7,19 @@
 class NSKalgo {
 public:
     // Secret key generation
-    static int secretkey(int p);
+    static long long int secretkey(long long int p);
 
     // Public key generation
-    static std::vector<int> publickey(int p, int s);
+    static std::vector<long long int> publickey(long long int p, long long int s);
 
     // Encryption
-    static std::vector<int> encrypt(const std::vector<int>& mess, const std::vector<int>& pubk, int p);
+    static std::vector<long long int> encrypt(const std::vector<long long int>& mess, const std::vector<long long int>& pubk, long long int p);
 
     // Decryption
-    static std::vector<int> decrypt(const std::vector<int>& cipher, int p, int s);
+    static std::vector<long long int> decrypt(const std::vector<long long int>& cipher, long long int p, long long int s);
 
     // Utility to dump contents of a vector (to replace the table dump)
-    static std::string dump(const std::vector<int>& vec);
+    static std::string dump(const std::vector<long long int>& vec);
 };
 
 #endif
